% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_experiment.R
\name{add_properties}
\alias{add_properties}
\title{Wrapper to jsPsych.data.addProperty}
\usage{
add_properties(...)
}
\arguments{
\item{...}{Name/value pairs}
}
\description{
Wrapper to jsPsych.data.addProperty
}
