% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_timeline.R
\name{build_timeline}
\alias{build_timeline}
\title{Build a timeline from trials}
\usage{
build_timeline(...)
}
\arguments{
\item{...}{trial objects to add to this timeline}
}
\description{
Build a timeline from trials
}
