% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_simple.R
\name{cue_image}
\alias{cue_image}
\title{Define an image cue}
\usage{
cue_image(stimulus, stimulus_height = NULL, stimulus_width = NULL,
  maintain_aspect_ratio = TRUE)
}
\arguments{
\item{stimulus}{Image file, specified using the resource() function}

\item{stimulus_height}{Image height in pixels (defaults to natural height)}

\item{stimulus_width}{Image width in pixels (defaults to natural width)}

\item{maintain_aspect_ratio}{Maintain aspect ratio (default = TRUE)}
}
\value{
An object of class xpt_cue
}
\description{
Define an image cue
}
