% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_condition.R
\name{data_condition}
\alias{data_condition}
\title{Create a javascript function that checks a data value}
\usage{
data_condition(expr, trials_back = 1)
}
\arguments{
\item{expr}{expression to be evaluated within the jsPsych data store}

\item{trials_back}{how many trials back?}
}
\description{
Create a javascript function that checks a data value
}
