% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_experiment.R
\name{define_resources}
\alias{define_resources}
\title{Construct a resource specification from paths}
\usage{
define_resources(from, audio = c(".mp3", ".wav", ".aif", ".mid"),
  video = c(".mp4", ".mpg", ".mov", ".wmv"), image = c(".jpg", ".png",
  ".bmp", ".svg", ".tiff"), script = c(".js"), style = c(".css"))
}
\arguments{
\item{from}{The paths to files/directories}

\item{audio}{File extensions assumed to be audio}

\item{video}{File extensions assumed to be video}

\item{image}{File extensions assumed to be images}

\item{script}{File extensions assumed to be scripts}

\item{style}{File extensions assumed to be stylesheets}
}
\description{
Construct a resource specification from paths
}
