% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_timeline.R
\name{display_while}
\alias{display_while}
\title{Loop this timeline}
\usage{
display_while(timeline, loop_function)
}
\arguments{
\item{timeline}{the timeline object}

\item{loop_function}{javascript function that returns true if loop repeats, false if terminates}
}
\description{
Loop this timeline
}
