% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_api.R
\name{download_googlecloud}
\alias{download_googlecloud}
\title{Download data from a jspsych experiment deployed on google cloud}
\usage{
download_googlecloud()
}
\description{
Download data from a jspsych experiment deployed on google cloud
}
\details{
This function currently does nothing
}
\seealso{
\code{\link{save_googlecloud}}, \code{\link{run_googlecloud}}
}
