% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_api.R
\name{download_webserver}
\alias{download_webserver}
\title{Download data from a jspsych experiment deployed on a webserver}
\usage{
download_webserver(ssh, keyfile = NULL, to = ".")
}
\arguments{
\item{ssh}{ssh connection string to a webserver}

\item{keyfile}{(optional) path to a ssh private key to log in to your webserver}

\item{to}{local folder to download the data to}
}
\description{
Download data from a jspsych experiment deployed on a webserver
}
\details{
This function assumes the default setup by run_webserver, so all it does
is download the folder /var/www/server_data
}
\seealso{
\code{\link{save_webserver}}, \code{\link{run_webserver}}
}
