% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_experiment.R
\name{experiment}
\alias{experiment}
\title{Make the experiment}
\usage{
experiment(timeline, path, resources = NULL, columns = NULL, ...)
}
\arguments{
\item{timeline}{tl}

\item{path}{p}

\item{resources}{r}

\item{columns}{columns}

\item{...}{pass to init}
}
\description{
Make the experiment
}
