% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_condition.R
\name{fn_data_condition}
\alias{fn_data_condition}
\title{Return a javascript function that checks a data value}
\usage{
fn_data_condition(expr, trials_back = 1)
}
\arguments{
\item{expr}{expression to be evaluated within the jsPsych data store}

\item{trials_back}{how many trials back?}
}
\description{
Return a javascript function that checks a data value
}
