% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomisation.R
\name{fn_sample}
\alias{fn_sample}
\title{Return a javascript function that samples from an array}
\usage{
fn_sample(x, size, replace = FALSE, weights = NULL)
}
\arguments{
\item{x}{A vector specifying the possible values}

\item{size}{The number of values to sample}

\item{replace}{Sample with replacement? (default = FALSE)}

\item{weights}{Probability of sampling each item (ignored if replace = FALSE)}
}
\value{
Javascript function
}
\description{
Return a javascript function that samples from an array
}
\details{
Loosely mirrors the sample function in base, but uses the jsPsych
randomisation functions to execute client side at runtime. At the moment it
does not work when x is a character vector, and always returns an array even
if the output is of length 1
}
