% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_locally.R
\name{fn_save_datastore}
\alias{fn_save_datastore}
\title{Return a javascript function to save data to Google datastore}
\usage{
fn_save_datastore()
}
\description{
Return a javascript function to save data to Google datastore
}
