% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullscreen.R
\name{fullscreen}
\alias{fullscreen}
\title{Toggle fullscreen mode in the browser}
\usage{
fullscreen(
  fullscreen_mode = TRUE,
 
    message = "<p>The experiment will switch to full screen mode when you press the button below</p>",
  button_label = "Continue",
  delay_after = 1000
)
}
\arguments{
\item{fullscreen_mode}{If TRUE, sets browser in full screen mode. Default: TRUE.}

\item{message}{This is the message that is displayed in the browser to inform of the switch to fullscreen mode.}

\item{button_label}{This is label of the button to acknowledge the switch.}

\item{delay_after}{Time period in ms after the switch to proceed with the following element in the timeline.}
}
\description{
The \code{fullscreen} function is used to toggle fullscreen mode in the browser.
}
