% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_variable}
\alias{insert_variable}
\title{Insert a timeline variable}
\usage{
insert_variable(name)
}
\arguments{
\item{name}{name of the variable to insert}
}
\description{
Insert a timeline variable
}
