% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_js.R
\name{jshelpers}
\alias{jshelpers}
\alias{js_code}
\alias{js_logical}
\alias{js_numeric}
\alias{js_string}
\alias{js_struct}
\alias{js_array}
\title{Manually specify the JavaScript object type}
\usage{
js_code(x)

js_logical(x)

js_numeric(x)

js_string(...)

js_struct(...)

js_array(...)
}
\arguments{
\item{x}{an object to be coerced}

\item{...}{objects to be concatenated and coerced}
}
\description{
Manually specify the JavaScript object type
}
