% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keycodes.R
\name{keycode}
\alias{keycode}
\title{Javascript key codes}
\usage{
keycode(key = NULL, code = NULL)
}
\arguments{
\item{key}{character vector specifying keynames (default = NULL)}

\item{code}{numeric vector specfiying keycodes (default = NULL)}
}
\value{
A numeric or character vector
}
\description{
Javascript key codes
}
\details{
This function provides a mapping between the human-readable
javascript \code{key} names and their corresponding numeric \code{code}
values. If both input arguments are \code{NULL}, it returns a named
numeric vector whose values correspond to the key codes and whose names
correspond to the key names. If \code{key} is specified the return value
is a vector with the corresponding numeric codes; whereas if \code{code}
is specified the output is a character vector containing the corresponding
key names. If neither argument is \code{NULL} the function throws an error.
}
