% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_survey.R
\name{question}
\alias{question}
\title{Create a survey question}
\usage{
question(cue, response, required = FALSE, name = NULL)
}
\arguments{
\item{cue}{the prompt for the question}

\item{response}{the response type for the question}

\item{required}{is a response to the question required?}

\item{name}{a convenient label for the question}
}
\value{
An object of class xpt_question
}
\description{
Create a survey question
}
