% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_survey_likert.R
\name{question_likert}
\alias{question_likert}
\title{Create a Likert question}
\usage{
question_likert(prompt, labels, required = FALSE, name = NULL)
}
\arguments{
\item{prompt}{the prompt for the question}

\item{labels}{the labels on the Likert scale}

\item{required}{is a response to the question required?}

\item{name}{a convenient label for the question}
}
\description{
Create a Likert question
}
