% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_survey_text.R
\name{question_text}
\alias{question_text}
\title{Create a free text response question}
\usage{
question_text(prompt, placeholder = "", rows = 1, columns = 40,
  name = NULL)
}
\arguments{
\item{prompt}{The prompt for the question}

\item{placeholder}{A string specifying the placeholder text}

\item{rows}{Number of rows spanned by the text box}

\item{columns}{Number of columns spanned by the text box}

\item{name}{A convenient label for the question}
}
\description{
Create a free text response question
}
