% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_survey.R
\name{response_freetext}
\alias{response_freetext}
\title{Specify a free text response}
\usage{
response_freetext(placeholder = "", rows = 1, columns = 40)
}
\arguments{
\item{placeholder}{String specifying the placeholder text}

\item{rows}{Number of rows spanned by the text box}

\item{columns}{Number of columns spanned by the text box}
}
\value{
An object of class xpt_response
}
\description{
Specify a free text response
}
