% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_simple.R
\name{response_key}
\alias{response_key}
\title{Specify a keyboard response}
\usage{
response_key(...)
}
\arguments{
\item{...}{Either single string "any" (default), "none", or characters specifying keys}
}
\value{
An object of class xpt_response
}
\description{
Specify a keyboard response
}
