% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_survey.R
\name{response_likert}
\alias{response_likert}
\title{Specify a Likert response}
\usage{
response_likert(labels, scale_width = NULL)
}
\arguments{
\item{labels}{character vector of response labels}

\item{scale_width}{width of the "slider" in pixels}
}
\value{
An object of class xpt_response
}
\description{
Specify a Likert response
}
