% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_survey.R
\name{response_pickone}
\alias{response_pickone}
\title{Specify a 'pick one' response}
\usage{
response_pickone(options, horizontal = FALSE)
}
\arguments{
\item{options}{character vector of options}

\item{horizontal}{should radio buttons be laid out horizontally?}
}
\value{
An object of class xpt_response
}
\description{
Specify a 'pick one' response
}
