% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_survey.R
\name{response_picksome}
\alias{response_picksome}
\title{Specify a 'pick some' response}
\usage{
response_picksome(options, horizontal = FALSE)
}
\arguments{
\item{options}{character vector of options}

\item{horizontal}{should check boxes be laid out horizontally?}
}
\value{
An object of class xpt_response
}
\description{
Specify a 'pick some' response
}
