% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_api.R
\name{run_appengine}
\alias{run_appengine}
\title{Deploy a jspsych experiment on google app engie}
\usage{
run_appengine(path, project_id)
}
\arguments{
\item{path}{path to the experiment folder}

\item{project_id}{the google app engine project id}
}
\description{
Deploy a jspsych experiment on google app engie
}
