% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_api.R
\name{run_googlecloud}
\alias{run_googlecloud}
\title{Deploy a jspsych experiment on google cloud}
\usage{
run_googlecloud(path, project_id)
}
\arguments{
\item{path}{path to the experiment folder}

\item{project_id}{the google app engine project id}
}
\description{
Deploy a jspsych experiment on google cloud
}
\details{
The purpose of the \code{run_googlecloud()} function is to make it
somewhat easier to deploy a jsPsych experiment to Google App Engine, so that
the experiment can run in the cloud rather than on the local machine.
The \code{path} argument
specifies the location of the experiment to be deployed, and should be the same
as the value of \code{path} that was used when calling
\code{\link{build_experiment}()}) to build the experiment originally. The
\code{project_id} is the name of the Google App Engine project that will host
the experiment.

At present, the functionality of \code{run_googlecloud()} is quite limited. All
it does is construct the appropriate command that you will need to enter at
the terminal. It does not execute that command, nor does it assist you in
creating the Google App Engine project itself (it is assumed that the user
already has a Google Cloud account and is authorised to deploy to the project)
}
\seealso{
\code{\link{save_googlecloud}}, \code{\link{build_experiment}}
}
