% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_api.R
\name{run_locally}
\alias{run_locally}
\title{Run a jspsych experiment}
\usage{
run_locally(path = ".", port = 8000)
}
\arguments{
\item{path}{the experiment directory}

\item{port}{port to use}
}
\description{
Run a jspsych experiment
}
