% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_api.R
\name{run_webserver}
\alias{run_webserver}
\title{Deploy a jspsych experiment to a webserver}
\usage{
run_webserver(
  path,
  experiment_folder = "experiment",
  ssh,
  keyfile = NULL,
  webserver_configured = FALSE
)
}
\arguments{
\item{path}{Path where the experiment is deployed}

\item{experiment_folder}{Experiment subfolder}

\item{ssh}{ssh connection string to a webserver}

\item{keyfile}{(optional) path to a ssh private key to log in to your webserver}

\item{webserver_configured}{set this to true when you've configured your webserver}
}
\description{
Deploy a jspsych experiment to a webserver
}
\details{
The purpose of the \code{run_webserver()} function is to make it
somewhat easier to deploy a jsPsych experiment to a webserver that you control,
so that the experiment can run in the cloud rather than on the local machine.

For this to work, you need to configure your webserver first. On a recent ubuntu image
(available from most cloud providers), you can install apache and php with
"\code{sudo apt install apache2 php}". You should probably also secure the connection
between the webserver and the participants with https. If you have configured a domain
name for your server, it is pretty simple to enable https via let's encrypt:
"\code{sudo add-apt-repository ppa:certbot/certbot}",
"\code{sudo apt install certbot python-certbot-apache}", and
"\code{sudo certbot --apache}"

Used together with \code{\link{save_webserver}}, this will set up a folder on the server
that apache can write to, and a little php script to receive said data
(as recommended on \href{https://www.jspsych.org/overview/data/#storing-data-permanently-as-a-file}{the jspsych website}).
}
\examples{
\dontrun{
build_experiment(..., on_finish = save_webserver())
run_webserver(ssh = "user@server.com", keyfile = "~/.ssh/id_rsa")
download_webserver(ssh = "user@server.com", keyfile = "~/.ssh/id_rsa")
}

}
\seealso{
\code{\link{save_webserver}}, \code{\link{build_experiment}}, \code{\link{download_webserver}}
}
