% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_webserver.R
\name{save_webserver}
\alias{save_webserver}
\title{Return a javascript function to save data via a script on the webserver}
\usage{
save_webserver()
}
\value{
Return a javascript function to save data via a script on the webserver
}
\description{
Return a javascript function to save data via a script on the webserver
}
\details{
The purpose of the \code{save_webserver()} is to return a
javascript function that, when called from within the jsPsych experiment,
will write the data to the server. This assumes that the experiment will
be run on a (php)script-enabled webserver. This way, you know the data will
never touch any other computer than the server you've presumably secured
and have data processing agreements in place for.
}
\seealso{
\code{\link{run_webserver}}, \code{\link{download_webserver}}
}
