% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_modify.R
\name{set_variables}
\alias{set_variables}
\title{Modify a timeline to set possible values for variables}
\usage{
set_variables(timeline, ...)
}
\arguments{
\item{timeline}{The timeline object}

\item{...}{A set of name/value pairs defining the timeline variables}
}
\value{
The modified timeline object
}
\description{
Modify a timeline to set possible values for variables
}
\details{
When creating an experiment, a common pattern is to create a
series of trials that are identical in every respect except for one thing
that varies across the trial (e.g., a collection of
\code{\link{trial_html_button_response}()} trials that are the same except
for the text that is displayed). A natural way to handle this in the
jsPsych framework is to create the trial in the usual fashion, except that
instead of specifying the \emph{value} that needs to be included in the
trial (e.g., the text itself) the code includes a reference to a
\emph{timeline variable}. Inserting the \emph{reference} to the variable
is the job of the \code{\link{insert_variable}()} function; \emph{attaching}
that variable to the timeline and specifying its possible values is the
job of \code{set_variables}. This is most easily explained by using
an example, as shown below.
}
\examples{
# create a template from which a series of trials can be built
template <- trial_html_button_response(stimulus = insert_variable("animal"))

# create a timeline with three trials, all using the same template
# but with a different value for the "animal" variable
timeline <- build_timeline(template) \%>\%
  set_variables(animal = c("cat", "dog", "pig"))

}
\seealso{
\code{\link{build_timeline}}, \code{\link{insert_variable}}
}
