% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_modify.R
\name{tl_add_parameters}
\alias{tl_add_parameters}
\title{Modify timeline to add parameters}
\usage{
tl_add_parameters(timeline, ...)
}
\arguments{
\item{timeline}{the timeline object}

\item{...}{name value pairs}
}
\description{
Modify timeline to add parameters
}
