% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_modify.R
\name{tl_add_variables}
\alias{tl_add_variables}
\title{Modify timeline to attach variables}
\usage{
tl_add_variables(timeline, ...)
}
\arguments{
\item{timeline}{the timeline object}

\item{...}{name value pairs}
}
\description{
Modify timeline to attach variables
}
