% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_modify.R
\name{tl_display_if}
\alias{tl_display_if}
\title{Modify timeline to execute if a condition is met}
\usage{
tl_display_if(timeline, conditional_function)
}
\arguments{
\item{timeline}{the timeline object}

\item{conditional_function}{javascript function that reutrns true to execute timeline, false to skip}
}
\description{
Modify timeline to execute if a condition is met
}
