% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_modify.R
\name{tl_display_while}
\alias{tl_display_while}
\title{Modify timeline to execute within a loop}
\usage{
tl_display_while(timeline, loop_function)
}
\arguments{
\item{timeline}{the timeline object}

\item{loop_function}{javascript function that returns true if loop repeats, false if terminates}
}
\description{
Modify timeline to execute within a loop
}
