% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_animation.R
\name{trial_animation}
\alias{trial_animation}
\title{Specify an animation trial}
\usage{
trial_animation(stimuli, frame_time = 250, frame_isi = 0,
  sequence_reps = 1, choices = respond_any_key(), prompt = NULL,
  post_trial_gap = 0, on_finish = NULL, on_load = NULL,
  data = NULL)
}
\arguments{
\item{stimuli}{A vector of paths to the image files}

\item{frame_time}{How long to display each image, in milliseconds}

\item{frame_isi}{How long is the gap between images, in milliseconds}

\item{sequence_reps}{How many times to repeat the sequence}

\item{choices}{A character vector of keycodes (either numeric values or the characters themselves). Alternatively, respond_any_key() and respond_no_key() can be used}

\item{prompt}{A string (may contain HTML) that will be displayed below the stimulus, intended as a reminder about the actions to take (e.g., which key to press).}

\item{post_trial_gap}{The gap in milliseconds between the current trial and the next trial. If NULL, there will be no gap.}

\item{on_finish}{A javascript callback function to execute when the trial finishes}

\item{on_load}{A javascript callback function to execute when the trial begins, before any loading has occurred}

\item{data}{An object containing additional data to store for the trial}
}
\value{
Functions with a \code{trial_} prefix always return a "trial" object.
A trial object is simply a list containing the input arguments, with
\code{NULL} elements removed. Logical values in the input (\code{TRUE} and
\code{FALSE}) are transformed to character vectors \code{"true"} and \code{"false"}
and are specified to be objects of class "json", ensuring that they will be
written to file as the javascript logicals, \code{true} and \code{false}.
}
\description{
The \code{trial_animation} function is used to display a
sequence of images at a fixed rate.
}
\details{
This function is used to specify an "animation" trial in a jsPsych experiment.
An animation trial displays a sequence of images at a fixed frame rate, and the sequence
can be looped a specified number of times. The participant is free to respond at any
point during the animation, and the time of the response is recorded.

The data recorded by an animation trial is as follows:

\itemize{
\item The \code{animation_sequence} value is an array encoded in JSON format. Each
element of the array is an object that represents a stimulus in the animation
sequence. Each object has a stimulus property, which is the image that was
displayed, and a time property, which is the time in ms, measured from when
the sequence began, that the stimulus was displayed.
\item The \code{responses} value is an array encoded in JSON format. Each element
of the array is an object representing a response given by the subject. Each
object has a stimulus property, indicating which image was displayed when the
key was pressed, an rt property, indicating the time of the key press relative
to the start of the animation, and a key_press property, indicating which key
was pressed.
}
}
