% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_categorise.R
\name{trial_categorize}
\alias{trial_categorize}
\title{Create a categorization trial}
\usage{
trial_categorize(cue, response, prompt = NULL,
  stimulus_duration = NULL, trial_duration = NULL,
  response_ends_trial = TRUE, ...)
}
\arguments{
\item{cue}{the cue type for the trial}

\item{response}{the response type for the question}

\item{prompt}{prompt to be displayed with the stimulus (can include HTML markup)}

\item{stimulus_duration}{how long to display stimulus in milliseconds? (default is to remain until response)}

\item{trial_duration}{how long to wait for response in milliseconds? (default is to wait indefinitely)}

\item{response_ends_trial}{does the trial end with response (default = TRUE) or continue until trial duration reached?}

\item{...}{arguments to be passed to trial()}
}
\value{
A trial object
}
\description{
Create a categorization trial
}
