% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_categorize_image.R
\name{trial_categorize_image}
\alias{trial_categorize_image}
\title{Specify a categorization trial with an image stimulus}
\usage{
trial_categorize_image(stimulus, key_answer, choices = respond_any_key(),
  text_answer = "", correct_text = "Correct",
  incorrect_text = "Wrong", prompt = NULL,
  force_correct_button_press = FALSE, show_stim_with_feedback = TRUE,
  show_feedback_on_timeout = FALSE,
  timeout_message = "Please respond faster", stimulus_duration = NULL,
  feedback_duration = 2000, trial_duration = NULL,
  post_trial_gap = 0, on_finish = NULL, on_load = NULL,
  data = NULL)
}
\arguments{
\item{stimulus}{The path to the image file to be displayed.}

\item{key_answer}{The numeric key code indicating the correct response}

\item{choices}{A character vector of keycodes (either numeric values or the characters themselves). Alternatively, respond_any_key() and respond_no_key() can be used}

\item{text_answer}{A label associated with the correct answer}

\item{correct_text}{Text to display when correct answer given ('\%ANS\%' substitutes text_answer)}

\item{incorrect_text}{Text to display when wrong answer given ('\%ANS\%' substitutes text_answer)}

\item{prompt}{A string (may contain HTML) that will be displayed below the stimulus, intended as a reminder about the actions to take (e.g., which key to press).}

\item{force_correct_button_press}{If TRUE the correct button must be pressed after feedback in order to advance}

\item{show_stim_with_feedback}{If TRUE the stimulus image will be displayed as part of the feedback. Otherwise only text is shown}

\item{show_feedback_on_timeout}{If TRUE the "wrong answer" feedback will be presented on timeout. If FALSE, a timeout message is shown}

\item{timeout_message}{The message to show on a timeout non-response}

\item{stimulus_duration}{How long to show the stimulus, in milliseconds. If NULL, then the stimulus will be shown until the subject makes a response}

\item{feedback_duration}{How long to show the feedback, in milliseconds}

\item{trial_duration}{How long to wait for a response before ending trial in milliseconds. If NULL, the trial will wait indefinitely. If no response is made before the deadline is reached, the response will be recorded as NULL.}

\item{post_trial_gap}{The gap in milliseconds between the current trial and the next trial. If NULL, there will be no gap.}

\item{on_finish}{A javascript callback function to execute when the trial finishes}

\item{on_load}{A javascript callback function to execute when the trial begins, before any loading has occurred}

\item{data}{An object containing additional data to store for the trial}
}
\value{
Functions with a \code{trial_} prefix always return a "trial" object.
A trial object is simply a list containing the input arguments, with
\code{NULL} elements removed. Logical values in the input (\code{TRUE} and
\code{FALSE}) are transformed to character vectors \code{"true"} and \code{"false"}
and are specified to be objects of class "json", ensuring that they will be
written to file as the javascript logicals, \code{true} and \code{false}.
}
\description{
The \code{trial_categorize_image} function is used to display
an image stimulus, collect a categorization response with the
keyboard, and provide feedback.
}
\details{
The data recorded by this trial is as follows:

\itemize{
\item The \code{stimulus} value is the path to the image file.
\item The \code{key_press}
value indicates which key the subject pressed. The value is the numeric key
code corresponding to the subject's response.
\item The \code{rt} value is the
response time in milliseconds for the subject to make a response. The time is
measured from when the stimulus first appears on the screen until the subject's
response.
\item The \code{correct} value is true if the subject got the correct answer,
false otherwise.
}

In addition, it records default variables that are recorded by all trials:

\itemize{
\item \code{trial_type} is a string that records the name of the plugin used to run the trial.
\item \code{trial_index} is a number that records the index of the current trial across the whole experiment.
\item \code{time_elapsed} counts the number of milliseconds since the start of the experiment when the trial ended.
\item \code{internal_node_id} is a string identifier for the current "node" in the timeline.
}
}
