% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_timeline.R
\name{trial_generic}
\alias{trial_generic}
\title{Adds an arbitrary trial to the timeline}
\usage{
trial_generic(type, ...)
}
\arguments{
\item{type}{the type of trial}

\item{...}{arguments passed to the trial plugin}
}
\description{
Adds an arbitrary trial to the timeline
}
