% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_html_slider_response.R
\name{trial_html_slider_response}
\alias{trial_html_slider_response}
\title{Specify an HTML trial with slider bar response}
\usage{
trial_html_slider_response(stimulus, labels = c("0\%", "25\%", "50\%",
  "75\%", "100\%"), button_label = "Continue", min = 0, max = 100,
  start = 50, step = 1, slider_width = NULL,
  require_movement = FALSE, prompt = NULL, stimulus_duration = NULL,
  trial_duration = NULL, response_ends_trial = TRUE,
  post_trial_gap = 0, on_finish = NULL, on_load = NULL,
  data = NULL)
}
\arguments{
\item{stimulus}{The HTML content to be displayed.}

\item{labels}{Labels displayed at equidistant locations on the slider.}

\item{button_label}{Label placed on the "continue" button}

\item{min}{Minimum value of the slider}

\item{max}{Maximum value of the slider}

\item{start}{Initial value of the slider}

\item{step}{Step size of the slider}

\item{slider_width}{Horizontal width of the slider (defaults to display width)}

\item{require_movement}{Does the user need to move the slider before clicking the continue button?}

\item{prompt}{A string (may contain HTML) that will be displayed below the stimulus, intended as a reminder about the actions to take (e.g., which key to press).}

\item{stimulus_duration}{How long to show the stimulus, in milliseconds. If NULL, then the stimulus will be shown until the subject makes a response}

\item{trial_duration}{How long to wait for a response before ending trial in milliseconds. If NULL, the trial will wait indefinitely. If no response is made before the deadline is reached, the response will be recorded as NULL.}

\item{response_ends_trial}{If TRUE, then the trial will end when a response is made (or the trial_duration expires). If FALSE, the trial continues until the deadline expires.}

\item{post_trial_gap}{The gap in milliseconds between the current trial and the next trial. If NULL, there will be no gap.}

\item{on_finish}{A javascript callback function to execute when the trial finishes}

\item{on_load}{A javascript callback function to execute when the trial begins, before any loading has occurred}

\item{data}{An object containing additional data to store for the trial}
}
\value{
Functions with a \code{trial_} prefix always return a "trial" object.
A trial object is simply a list containing the input arguments, with
\code{NULL} elements removed. Logical values in the input (\code{TRUE} and
\code{FALSE}) are transformed to character vectors \code{"true"} and \code{"false"}
and are specified to be objects of class "json", ensuring that they will be
written to file as the javascript logicals, \code{true} and \code{false}.
}
\description{
The \code{trial_html_slider_response} function is used to display
an HTML stimulus and collect a response using a slider bar.
}
\details{
This plugin displays an HTML stimulus and records responses generated by a slider. The
stimulus can be displayed until a response is given, or for a pre-determined amount of time. The
trial can be ended automatically if the subject has failed to respond within a fixed length of time.

The data recorded by this trial is as follows:

\itemize{
\item The \code{rt} value is the response time in milliseconds taken for the
user to make a response. The time is measured from when the stimulus first
appears on the screen until the response.
\item The \code{response} is the numeric value of the slider bar.
\item The \code{stimulus} variable records the HTML content that was
displayed on this trial.
}

In addition, it records default variables that are recorded by all trials:

\itemize{
\item \code{trial_type} is a string that records the name of the plugin used to run the trial.
\item \code{trial_index} is a number that records the index of the current trial across the whole experiment.
\item \code{time_elapsed} counts the number of milliseconds since the start of the experiment when the trial ended.
\item \code{internal_node_id} is a string identifier for the current "node" in the timeline.
}
}
