% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_image_keyboard_response.R
\name{trial_image_keyboard_response}
\alias{trial_image_keyboard_response}
\title{Displays an image stimulus and records responses generated with the keyboard}
\usage{
trial_image_keyboard_response(stimulus, stimulus_height = NULL,
  stimulus_width = NULL, maintain_aspect_ratio = TRUE,
  choices = any_key(), prompt = NULL, stimulus_duration = NULL,
  trial_duration = NULL, response_ends_trial = TRUE,
  post_trial_gap = 0, on_finish = NULL, on_load = NULL,
  data = NULL)
}
\arguments{
\item{stimulus}{The path of the image file to be displayed.}

\item{stimulus_height}{Set the height of the image in pixels. If NULL, then the image will display at its natural height.}

\item{stimulus_width}{Set the width of the image in pixels. If NULL, then the image will display at its natural width.}

\item{maintain_aspect_ratio}{If setting only the width or only the height and this parameter is TRUE, then the other dimension will be scaled to maintain the image's aspect ratio.}

\item{choices}{A character vector of keycodes (either numeric values or the characters themselves). Alternatively, any_key() and no_key() can be used}

\item{prompt}{A string (may contain HTML) that will be displayed below the stimulus, intended as a reminder about the actions to take (e.g., which key to press).}

\item{stimulus_duration}{How long to show the stimulus, in milliseconds. If NULL, then the stimulus will be shown until the subject makes a response}

\item{trial_duration}{How long to wait for a response before ending trial in milliseconds. If NULL, the trial will wait indefinitely. If no response is made before the deadline is reached, the response will be recorded as NULL.}

\item{response_ends_trial}{If TRUE, then the trial will end when a response is made (or the trial_duration expires). If FALSE, the trial continues until the deadline expires.}

\item{post_trial_gap}{The gap in milliseconds between the current trial and the next trial. If NULL, there will be no gap.}

\item{on_finish}{A javascript callback function to execute when the trial finishes}

\item{on_load}{A javascript callback function to execute when the trial begins, before any loading has occurred}

\item{data}{An object containing additional data to store for the trial}
}
\description{
Displays an image stimulus and records responses generated with the keyboard
}
\details{
This plugin displays an image and records responses generated by a key press. The
stimulus can be displayed until a response is given, or for a pre-determined amount of time. The
trial can be ended automatically if the subject has failed to respond within a fixed length of time.

In addition to the default data collected by all plugins, this plugin collects the following data for
each trial: the \code{rt} value is the response time in milliseconds taken for the user to make a
response. The time is measured from when the stimulus first appears on the screen until the response;
\code{key_press} is the numeric key code corresponding to the response; \code{stimulus} The HTML content
that was displayed on the screen.
}
