% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_instructions.R
\name{trial_instructions}
\alias{trial_instructions}
\title{Show one or more pages of instructions to the user}
\usage{
trial_instructions(pages, key_forward = "rightarrow",
  key_backward = "leftarrow", allow_backward = TRUE,
  allow_keys = TRUE, show_clickable_nav = FALSE,
  button_label_previous = "Previous", button_label_next = "Next",
  post_trial_gap = 0, on_finish = NULL, on_load = NULL,
  data = NULL)
}
\arguments{
\item{pages}{Character vector. Each element should be an HTML-formatted string specifying a page}

\item{key_forward}{This is the key that the subject can press in order to advance to the next page, specified as their numeric key code or as characters}

\item{key_backward}{This is the key that the subject can press in order to return to the previous page.}

\item{allow_backward}{If TRUE, participants can navigate backwards}

\item{allow_keys}{If TRUE, participants can use keyboard keys to navigate}

\item{show_clickable_nav}{If TRUE, buttons will be shown to allow navigation}

\item{button_label_previous}{Text on the "previous" button}

\item{button_label_next}{Thext on the "next" button}

\item{post_trial_gap}{The gap in milliseconds between the current trial and the next trial. If NULL, there will be no gap.}

\item{on_finish}{A javascript callback function to execute when the trial finishes}

\item{on_load}{A javascript callback function to execute when the trial begins, before any loading has occurred}

\item{data}{An object containing additional data to store for the trial}
}
\description{
Show one or more pages of instructions to the user
}
\details{
This plugin is for showing instructions to the subject. It allows
subjects to navigate through multiple pages of instructions at their own pace,
recording how long the subject spends on each page. Navigation can be done
using the mouse or keyboard. Subjects can be allowed to navigate forwards
and backwards through pages, if desired.

In addition to the default data collected by all plugins, this
plugin collects the following data for each trial:

The \code{view_history} value is a JSON string containing the order of pages
the subject viewed (including when the subject returned to previous pages)
and the time spent viewing each page. The \code{rt} value is the response time
in milliseconds for the subject to view all of the pages.
}
