% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_survey.R
\name{trial_survey}
\alias{trial_survey}
\title{Create a survey trial}
\usage{
trial_survey(questions, preamble = "",
  randomize_question_order = FALSE, button_label = "Continue")
}
\arguments{
\item{questions}{A question or list of questions}

\item{preamble}{Text to appear above the questions}

\item{randomize_question_order}{Should order be randomised?}

\item{button_label}{Text for the continue button}
}
\value{
A trial object
}
\description{
Create a survey trial
}
