% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_survey_likert.R
\name{trial_survey_likert}
\alias{trial_survey_likert}
\title{A survey page with Likert scale items}
\usage{
trial_survey_likert(questions, preamble = "", scale_width = NULL,
  randomize_question_order = FALSE, button_label = "Continue",
  post_trial_gap = 0, on_finish = NULL, on_load = NULL,
  data = NULL)
}
\arguments{
\item{questions}{A question or list of questions}

\item{preamble}{Text to appear above the questions}

\item{scale_width}{Width of the scale in pixels (NULL is the display width)}

\item{randomize_question_order}{Should order be randomised?}

\item{button_label}{Text for the continue button}

\item{post_trial_gap}{The gap in milliseconds between the current trial and the next trial. If NULL, there will be no gap.}

\item{on_finish}{A javascript callback function to execute when the trial finishes}

\item{on_load}{A javascript callback function to execute when the trial begins, before any loading has occurred}

\item{data}{An object containing additional data to store for the trial}
}
\description{
A survey page with Likert scale items
}
