% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_survey_multi_choice.R
\name{trial_survey_multi_choice}
\alias{trial_survey_multi_choice}
\title{Specify a survey page with multiple choice items}
\usage{
trial_survey_multi_choice(questions, preamble = "",
  randomize_question_order = FALSE, button_label = "Continue",
  required_message = "You must choose at least one response for this question",
  post_trial_gap = 0, on_finish = NULL, on_load = NULL,
  data = NULL)
}
\arguments{
\item{questions}{A question or list of questions}

\item{preamble}{Text to appear above the questions}

\item{randomize_question_order}{Should order be randomised?}

\item{button_label}{Text for the continue button}

\item{required_message}{Message to display if required response is not given.}

\item{post_trial_gap}{The gap in milliseconds between the current trial and the next trial. If NULL, there will be no gap.}

\item{on_finish}{A javascript callback function to execute when the trial finishes}

\item{on_load}{A javascript callback function to execute when the trial begins, before any loading has occurred}

\item{data}{An object containing additional data to store for the trial}
}
\value{
Functions with a \code{trial_} prefix always return a "trial" object.
A trial object is simply a list containing the input arguments, with
\code{NULL} elements removed. Logical values in the input (\code{TRUE} and
\code{FALSE}) are transformed to character vectors \code{"true"} and \code{"false"}
and are specified to be objects of class "json", ensuring that they will be
written to file as the javascript logicals, \code{true} and \code{false}.
}
\description{
The \code{trial_survey_multi_choice} function is used to display a
survey page with one or more multiple choice questions
}
\details{
The \code{trial_survey_multi_choice} function creates a trial that displays
a set of questions with multiple choice responses. The \code{questions} argument is
used to specify the questions themselves, and can either consist of a single
question object generated by \code{\link{question_multi}} or a list of such objects.
The multiple choice items are laid out as a set of radio buttons with labels specified
when calling \code{\link{question_multi}}, and the participant can select only one
possible response.


The data recorded by this trial is as follows:
\itemize{
\item The \code{responses} value is a
an array containing all selected choices in JSON format for each question. The
encoded object will have a separate variable for the response to each question,
with the first question in the trial being recorded in Q0, the second in Q1, and
so on. The responses are recorded as the name of the option label. If the
\code{name} parameter is defined for the question, then the response will use
the value of \code{name} as the key for the response in the responses object.

\item The \code{rt} value is the response time in milliseconds for the subject to make
a response. The time is measured from when the questions first appear on the
screen until the subject's response.

\item The \code{question_order} value is a string in JSON format containing an array
with the order of questions. For example [2,0,1] would indicate that the first
question was trial.questions[2] (the third item in the questions parameter), the
second question was trial.questions[0], and the final question was trial.questions[1].
}

In addition, it records default variables that are recorded by all trials:

\itemize{
\item \code{trial_type} is a string that records the name of the plugin used to run the trial.
\item \code{trial_index} is a number that records the index of the current trial across the whole experiment.
\item \code{time_elapsed} counts the number of milliseconds since the start of the experiment when the trial ended.
\item \code{internal_node_id} is a string identifier for the current "node" in the timeline.
}
}
\seealso{
Survey page trials are constructed using the \code{\link{trial_survey_text}},
\code{\link{trial_survey_likert}}, \code{\link{trial_survey_multi_choice}} and
\code{\link{trial_survey_multi_select}} functions. Individual questions for survey
trials can be specified using \code{\link{question_text}},
\code{\link{question_likert}} and \code{\link{question_multi}}.
}
