% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_timeline.R
\name{with_condition}
\alias{with_condition}
\title{Conditionally execute this timeline}
\usage{
with_condition(timeline, conditional_function)
}
\arguments{
\item{timeline}{the timeline object}

\item{conditional_function}{javascript function that reutrns true to execute timeline, false to skip}
}
\description{
Conditionally execute this timeline
}
