% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_timeline.R
\name{with_parameters}
\alias{with_parameters}
\title{Attach a timeline variable to timeline object}
\usage{
with_parameters(timeline, ...)
}
\arguments{
\item{timeline}{the timeline object}

\item{...}{name value pairs}
}
\description{
Attach a timeline variable to timeline object
}
