% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomisation.R
\name{fn_sample}
\alias{fn_sample}
\title{Return a javascript function that samples from an array}
\usage{
fn_sample(x, size, replace = FALSE, weights = NULL)
}
\arguments{
\item{x}{A vector specifying the possible values}

\item{size}{The number of values to sample}

\item{replace}{Sample with replacement? (default = FALSE)}

\item{weights}{Probability of sampling each item (ignored if replace = FALSE)}
}
\value{
Returns a javascript function that samples from an array of values
}
\description{
Return a javascript function that samples from an array
}
\details{
The \code{fn_sample()} is used to return a function that, when called
from within a jsPsych experiment, will mirror the behaviour of the \code{sample()}
function from the base package using th jsPsych randomisation functions to
at runtime. The input argument \code{x} specifies the set of values from which
samples should be drawn, and the \code{size} argument specifies the number of
samples to be drawn. When \code{replace = TRUE} items are sampled with
replacement, and when \code{replace = FALSE} items are sampled without
replacement. When sampling with replacement, the \code{weights} argument can
be used to specify unequal sampling probabilities.

The current implementation is limited. It does not work when \code{x} is a
character vector, for example. Note also that the value returned within
the jsPsych experiment is always an array (not a scalar), even when
\code{size = 1}.
}
