% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_javascript}
\alias{insert_javascript}
\title{Insert input as raw javascript}
\usage{
insert_javascript(string)
}
\arguments{
\item{string}{a string to be interpreted as javascript code}
}
\description{
Insert input as raw javascript
}
