% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_property}
\alias{insert_property}
\title{Insert a property to the jsPsych data store}
\usage{
insert_property(...)
}
\arguments{
\item{...}{Name/value pairs}
}
\description{
Insert a property to the jsPsych data store
}
