% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert.R
\name{insert_resource}
\alias{insert_resource}
\title{Insert input as path to a resource file}
\usage{
insert_resource(file)
}
\arguments{
\item{file}{path}
}
\description{
Insert input as path to a resource file
}
