% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_survey_multi_choice.R
\name{question_multi}
\alias{question_multi}
\title{Create a multiple choice/select question}
\usage{
question_multi(prompt, options, horizontal = FALSE, required = FALSE,
  name = NULL)
}
\arguments{
\item{prompt}{the prompt for the question}

\item{options}{character vector of options}

\item{horizontal}{should radio buttons be laid out horizontally?}

\item{required}{is a response to the question required?}

\item{name}{a convenient label for the question}
}
\description{
Create a multiple choice/select question
}
\seealso{
Survey page trials are constructed using the \code{\link{trial_survey_text}},
\code{\link{trial_survey_likert}}, \code{\link{trial_survey_multi_choice}} and
\code{\link{trial_survey_multi_select}} functions. Individual questions for survey
trials can be specified using \code{\link{question_text}},
\code{\link{question_likert}} and \code{\link{question_multi}}.
}
