% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_api.R
\name{run_locally}
\alias{run_locally}
\title{Deploy a jspsych experiment locally}
\usage{
run_locally(path, port = 8000)
}
\arguments{
\item{path}{Path to the experiment folder}

\item{port}{The port to use}
}
\description{
Deploy a jspsych experiment locally
}
\details{
The purpose of the \code{run_locally()} function is to start an
R server running (using the plumber package) that will
serve the experiment from the local machine. The \code{path} argument
specifies the location of the experiment to be run, and should be the same
as the value of \code{path} that was used when calling
\code{\link{build_experiment}()}) to build the experiment originally. The
\code{port} is the numeric value of the port on which the experiment is served.
Once \code{run_locally()} has been called, a browser window should open
showing the relevant page.

There are two reasons to deploy a local experiment using \code{run_locally()}
rather than simply opening the relevant "index.html" file in the browser. The
first is for the purpose of saving data. For security reasons, browsers do not
generally permit client-side javascript (e.g., the code that runs the jsPsych
experiment) to save files to arbitrary locations. For this reason writing the
data to file is the job of the R server, not the javascript code that
is running through the browser. In other words, if the experiment is deployed
locally using the \code{run_locally()} function, then the \code{\link{save_locally}()}
function that used to record data locally will work properly. If, however, the
"index.html" file is opened without starting the R server, data will not be
saved to file.

The second reason for using \code{run_locally()} is that it opens up the possibility
that an experiment could use server-side R code at runtime. At the moment jaysire
does not have any functionality to do so, but in principle there is nothing
preventing the R server from playing a more active role when the experiment is
running, and future versions of the package may develope this functionality
further.
}
\seealso{
\code{\link{save_locally}}, \code{\link{build_experiment}}
}
