% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_categorize_animation.R
\name{trial_categorize_animation}
\alias{trial_categorize_animation}
\title{Displays animation, records responses with the keyboard and gives feedback}
\usage{
trial_categorize_animation(stimuli, key_answer,
  choices = respond_any_key(), text_answer = "",
  correct_text = "Correct", incorrect_text = "Wrong",
  frame_time = 500, sequence_reps = 1,
  allow_response_before_complete = FALSE, prompt = NULL,
  feedback_duration = 2000, post_trial_gap = 0, on_finish = NULL,
  on_load = NULL, data = NULL)
}
\arguments{
\item{stimuli}{Character vector of paths to image files}

\item{key_answer}{The numeric key code indicating the correct response}

\item{choices}{A character vector of keycodes (either numeric values or the characters themselves). Alternatively, respond_any_key() and respond_no_key() can be used}

\item{text_answer}{A label associated with the correct answer}

\item{correct_text}{Text to display when correct answer given ('\%ANS\%' substitutes text_answer)}

\item{incorrect_text}{Text to display when wrong answer given ('\%ANS\%' substitutes text_answer)}

\item{frame_time}{How long to display each image, in milliseconds}

\item{sequence_reps}{How many times to show the entire sequence}

\item{allow_response_before_complete}{If TRUE the user can respond before the animation sequence finishes}

\item{prompt}{A string (may contain HTML) that will be displayed below the stimulus, intended as a reminder about the actions to take (e.g., which key to press).}

\item{feedback_duration}{How long to show the feedback, in milliseconds}

\item{post_trial_gap}{The gap in milliseconds between the current trial and the next trial. If NULL, there will be no gap.}

\item{on_finish}{A javascript callback function to execute when the trial finishes}

\item{on_load}{A javascript callback function to execute when the trial begins, before any loading has occurred}

\item{data}{An object containing additional data to store for the trial}
}
\description{
Displays animation, records responses with the keyboard and gives feedback
}
\details{
In addition to the default data collected by all plugins, this
plugin collects the following data for each trial:

The \code{stimulus} value is a JSON encoded representation of the array of
stimuli displayed in the trial. The \code{key_press}
value indicates which key the subject pressed. The value is the numeric key
code corresponding to the subject's response. The \code{rt} value is the
response time in milliseconds for the subject to make a response. The time is
measured from when the stimulus first appears on the screen until the subject's
response. The \code{correct} value is true if the subject got the correct answer,
false otherwise.
}
