% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_generic.R
\name{trial_generic}
\alias{trial_generic}
\title{Adds an arbitrary trial to the timeline}
\usage{
trial_generic(type, ...)
}
\arguments{
\item{type}{the type of trial}

\item{...}{arguments passed to the trial plugin}
}
\value{
Functions with a \code{trial_} prefix always return a "trial" object.
A trial object is simply a list containing the input arguments, with
\code{NULL} elements removed. Logical values in the input (\code{TRUE} and
\code{FALSE}) are transformed to character vectors \code{"true"} and \code{"false"}
and are specified to be objects of class "json", ensuring that they will be
written to file as the javascript logicals, \code{true} and \code{false}.
}
\description{
Adds an arbitrary trial to the timeline
}
